// dotnet/openapi/model/applicantModelServiceResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ApplicantModel } from './applicantModel';


export interface ApplicantModelServiceResponse { 
    message?: string | null;
    statusCode?: number;
    data?: ApplicantModel;
}