// dotnet/openapi/model/createApplicantCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateApplicantCommand { 
    fullName?: string | null;
    email?: string | null;
    phone?: string | null;
    resumeUrl?: string | null;
    skills?: string | null;
    experience?: number;
    jobId?: string;
}