// dotnet/openapi/model/updateApplicantCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateApplicantCommand { 
    fullName?: string | null;
    email?: string | null;
    phone?: string | null;
    resumeUrl?: string | null;
    skills?: string | null;
    experience?: number;
    jobId?: string;
    id?: string;
}